Move CanonicalAddress to persistence package
