+++ src/main/java/uk/gov/di/ipv/cri/common/library/persistence/item/personidentity/PersonIdentityItem.java
@@ -2,7 +2,7 @@ package uk.gov.di.ipv.cri.common.library.persistence.item.personidentity;
 
 import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbBean;
 import software.amazon.awssdk.enhanced.dynamodb.mapper.annotations.DynamoDbPartitionKey;
-import uk.gov.di.ipv.cri.common.library.domain.CanonicalAddress;
+import uk.gov.di.ipv.cri.common.library.persistence.item.CanonicalAddress;
 
 import java.util.List;
 import java.util.UUID;
