+++ src/test/java/uk/gov/di/ipv/cri/common/library/service/PersonIdentityMapperTest.java
@@ -4,7 +4,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
-import uk.gov.di.ipv.cri.common.library.domain.CanonicalAddress;
 import uk.gov.di.ipv.cri.common.library.domain.personidentity.PersonAddress;
 import uk.gov.di.ipv.cri.common.library.domain.personidentity.PersonAddressType;
 import uk.gov.di.ipv.cri.common.library.domain.personidentity.PersonIdentity;
@@ -13,6 +12,7 @@ import uk.gov.di.ipv.cri.common.library.domain.sharedclaims.BirthDate;
 import uk.gov.di.ipv.cri.common.library.domain.sharedclaims.Name;
 import uk.gov.di.ipv.cri.common.library.domain.sharedclaims.NamePart;
 import uk.gov.di.ipv.cri.common.library.domain.sharedclaims.SharedClaims;
+import uk.gov.di.ipv.cri.common.library.persistence.item.CanonicalAddress;
 import uk.gov.di.ipv.cri.common.library.persistence.item.personidentity.PersonIdentityDateOfBirth;
 import uk.gov.di.ipv.cri.common.library.persistence.item.personidentity.PersonIdentityItem;
 import uk.gov.di.ipv.cri.common.library.persistence.item.personidentity.PersonIdentityName;
